VIRO-595: animation refactor (bridge)
Refactor the bridge to use the new animation
execution API in the renderer. The VRTAnimatedComponent
is simplified so it's just a glue-layer between react
and the renderer. The VRTAnimationManager is
simplified in that it simply parses animations.

Chained animations are also more flexible now; they
aren't restricted to just using other named animations,
you can create chains of chains and directly embed
animation dictionaries in chains.


Former-commit-id: 758f08f80de8bbfabeb77863fdab9099e140f0d7
Former-commit-id: 74098b8301f7df071d410041b11ae011b975c79c