VIRO-817: allow multiple cameras per "node"
The renderer restricts the system to one camera
per node, but React allowed multiple cameras to
be placed in a node. The ideal situation is we can
place multiple cameras, but only the 'active' camera
in the node is displayed.

Enable this by making cameras extend from nodes
themselves, similar to how we handle 3D objects,
text, and other controls.

This also simplifies the camera logic in the bridge,
in that we don't have to grab the camera's superview
to set the point of view. It also enables the user to
set a rotation on the node, which determines the
rotation of the camera.

Note that the position property impacts the camera's
position, not the node's position. This is to ensure that
the focalPoint remains relative to the camera's parent
node, and not relative to the camera's position.


Former-commit-id: 3a764bea73dbeacc35d689b99ccbb943ef330fd9
Former-commit-id: 808d960d8045f4cdf4edf91a39c74e64f59b0ada