VIRO-877: fix tear-down so sounds remove
The fix for VIRO-818 (changing tear-downs so they
only occur when dropped AND removed) broke
tear-downs for elements that should be destroyed
when they're simply removed.

To simplify all this we just tear down whenever a
component is removed. This makes the sound stop
playing upon removal.

It's unclear when react will remove a view and not
expect it to be torn down. We're not sure if it attempts
to ever re-use these views.


Former-commit-id: 7090c5689c56df1e4e092cc657ebe0d814d4827d
Former-commit-id: 8fb0c13c21f1be27fc0f0eb213e41d7990dba056 [formerly 09dd240d0bb63d6f975755f6a386cce91812fa5b]
Former-commit-id: 71e13aeeeae88f12fe8e21d4d2aad7793ab3b27c