VIRO-919: crash when launching if views removed
When launching the TestBed app, if you're in portrait
mode 2 views get added (the VR view and the first
scene). Then the app switches to landscape so those
two views get removed, and added again. In this
interval, if the render context is set, we get a crash.

Add an additional check to ensure this doesn't happen.

Why the views are added and removed is still an unknown,
but this commit is meant primarily unblock the testbed app.


Former-commit-id: 83a05b11e5ef8c3b3f5a5e522276fac9115a487f
Former-commit-id: 8a33d52b0fd88aeef1aa8c498c1f8ad497b8175f [formerly 38b7e9aee60aadd670fe4165488bd23274c63ad0]
Former-commit-id: a8afadc2215dff47a4ebab59a0a2a8febf65c749