VIRO-799: button tap and gaze source fixes
If tap or gaze source was missing we would get an
error. Fix this so instead if gaze source is missing
we just use the normal image for gaze. If tap source
is missing we use the gaze source, and if there's
no gaze source, we use the normal source.


Former-commit-id: b18b946d02ecec6c8409b20c0803794052cef431
Former-commit-id: 116e90e7ef25ea413b1128b34b99f0393d069413 [formerly 3be48ff03ed5871cfa2d73967069763c1deece80]
Former-commit-id: 1f32df8af6e48fe7c26de8d1ceb02c80b1e9c39f