VIRO-947: onTearDown when dropped and detached
Fix onTearDown so that we only tear views down when
they are both dropped and detached (meaning they have
no children and no parent).

Tested and this seems to work: complex view trees are
torn down, but only after React is done processing them.

Still crashes on animations, which continue to run on
town-down components.


Former-commit-id: 9c4b3298c941f895e934227be8ebde08d8c336a1
Former-commit-id: 4893bcf9935ada7365b67c63fcbdc4f170cb3674 [formerly 1ee006db99611aa28e6e99c68fe84afe2f6a33ad]
Former-commit-id: 6a14dc2766c95873bc05722f72557439c772ad5b