VIRO-1052 NPE Crash in Testbed MediaPlayer/360 WeWork
- We noticed that AnimatedComponent didn't implement removeViewAt() which
is what React Native uses to remove views, but that's causing us issues
now because for some reason the Viro Media App is restarting the scene 3 times
and this is now causing us some issues. Removing this for now since things
"worked" before, we are probably leaking something here though.


Former-commit-id: 9b78161b08aab2c339b5482063ca113573bd1017
Former-commit-id: 79caf38071d60a5f1c22c636039abf27cfea73c9 [formerly 5150ecd762747cce7c239500fdd21202c7096345]
Former-commit-id: cea9935db475e7da4273e13376936f48e3496322