VIRO-918 [bridge]: preserve rendering tree when animate out
The second issue when animating out was the Android
React Native system was digging into the bridge tree
and explicitly removing all children from their parents.

With this change, when we detect a scene is being removed,
we immediately force a cascade tear down. This has the
effect of preserving the rendering tree, since removeViewAt
calls all check for tear down status before processing the
removal. Two reasons why this is ok:

1. Scenes are always meant to be town down anyway
after they’re removed from the SceneNavigator.

2. The entire rendering tree still cleans itself up (memory-wise)
since it’s detached from the scene controller and the scene
navigator. This is all detailed in the comment added to
SceneNavigator.java.


Former-commit-id: 4f34e111b8cb7e357cf0f2d02d8c966845f2f519
Former-commit-id: 368e74c627c27a511a36bf7c9b43804c51abc8df [formerly dc2efda8fb5c2a3572c40825ad03dcdb991b5623]
Former-commit-id: ff6e8d7671be509131ca770d28acdbf1982b501b