VIRO-1314: factor out managed animation
Pull out the animation management from animated
component into ManagedAnimation.

Also updates the renderer and restores compatibility
with the latest version.

Note this introduces a regression whereby animation
properties that are changed are not reflected in the
new animation (matching a regression on iOS), which
will be addressed as a bug fix.


Former-commit-id: 804cbd2e174e7483e60d7cd47b8a44b2c0c6ed06
Former-commit-id: 20666b5828ea499e9c02578e57a6349e79543568 [formerly d2ec23cef71a767c5d261509a00195c6c7e508ce]
Former-commit-id: 659676b6ecd0e98096443125153a1f6a4b08a934