VIRO-2104: ViroPolylineTest Crashes onLoad
The camera was crashing when we attempt to build a position Vector
with a null float array that occurs when the position is not set.
As such the fix is to pre-initialize the positional value, and have
the camera use the right position property.

Finally, also set missing nativeOnly props for ViroPolyline.


Former-commit-id: 22b74d1ad1cc2a64dfd5cbd1b2b7e0b1be1ce331
Former-commit-id: b9e382732a95e074d5b18cb4c991c85328e285fd [formerly 8a30a23868a7e3f617b925fec09d83933e4693ea]
Former-commit-id: 6ee49bf0a14c5b16d85f34f85b3c6ad2a1de14a2