VIRO-2260: Setting ViroSkyBox background is not set per-portal
Setting the background on a ViroSkyBox always sets the scene's background,
and is not portal specific. The fix is to emulate what we do for 360Backgrounds,
where we treat the skyBox as a node and utilize the node's getParentPortalScene
to identify the appropriate portalscene to apply the background onto, hence making
the skybox portal-specific.


Former-commit-id: 4c218230360e42aecd76e331967621073c899627
Former-commit-id: cd7bbffd136b90fcdfd54fa538200369f42649b3 [formerly 6b3a158ed46c1a84ae2187517a4ebc40d7afc02a]
Former-commit-id: 1b3d7d2f8414aae1ff70b0a82211e5a970cf622d