VIRO-2486: do not force landscape in 360 mode (bridge)
Change VRTSceneNavigator so it can use the new
forceLandscape boolean on VROViewControllerGVR
to make it such that we do *not* force landscape
mode when using 360.

Unfortunately because the VROViewControllerGVR
needs this information right after construction, we
have to defer the creation of the
VROViewControllerGVR until setVrMode has been
called. To ensure this always works, we lazily load
the VROViewControllerGVR using a new method.


Former-commit-id: 9402fef2c9f99fb23815e50931574c4fc6feaffc
Former-commit-id: 4ea5ed22812b0d2fa719f79e6e3b50471a1882a8 [formerly 872f3b821e8c31610af90b0eabdbb4549eb754ae]
Former-commit-id: 34d278acf2705cbb6a173f7237e73cbd28d172b8