VIRO-3738: ARCore anchors in React
Utilize ARCore anchors in the React bridge. Whenever
a Node is added to a VRTARScene, we attempt to create
an anchor to intermediate that Node’s position in the
scene. In other words, if we’re adding a Node called
Node A, instead of having:

ARScene
   — Node A

we have:

ARScene
   — Anchor A
       — Node A

We do this behind the scenes via the AnchorAttempt
class in VRTNode. A few notes:

1. If a Node is removed, we detach the anchor
2. If a Node is moved, we detach the anchor and move
    the Node to the anchor’s position
3. If a Node’s position is changed, we either move the
    Node relative to the anchor (if the distance is small),
    or “re-anchor” the Node entirely, by floating the Node,
    detaching its previous anchor, and creating a new
    anchor.
4. If anchoring fails, we try again 3 times. If all attempts
    fail, we simply float the Node.


Former-commit-id: c5b026d2f8d6c98109a6a5786d34944d92dd96de
Former-commit-id: 54346f5a3723006beed1dc06bc7782ba1dd82aa7 [formerly ae2068ff715234d505846b24c5fa442609140cf3]
Former-commit-id: a43c77b634dce6a923a4d2c758460bf7b1fed822