VIRO-4038: duration override in React API
Adds ability to override the duration of a skeletal
or keyframe animation. Because this only applies
to those types of animations, we only include this
property in Viro3DObject, not in other components.
Other components should use the animation’s set
duration; or we can consider making duration
override universal (e.g. also make it work for
registered animations).


Former-commit-id: 3f4ab05342ddab0ee7a32c3d76d20cedba6e37ad
Former-commit-id: 183097854f5b5958880fe2c8b7d60f93fb0f3958 [formerly 6c17ff2fc149477308b61063396f16aaa5af4f52]
Former-commit-id: 924195316dfc753416d47f6fc6f5445f562d0152